requirements.txt
flask
snowflake-snowpark-python